Extract tests for isWeakSet to own file
